reorganized the project structure
